project (3)/project/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
requests==2.31.0
sqlite3